Moving courier logic in useCourierStore.
