+++ src/pages/admin/EditCourier.jsx
@@ -1,8 +1,8 @@
+import { useEffect } from "react";
 import { useParams, useNavigate } from "react-router-dom";
 import { useCourierStore } from "../../store/useCourierStore";
 import CourierForm from "../../components/admin/CourierForm";
 import { useTranslation } from "react-i18next";
-import { toCourierPayload } from "../../services/courierService";
 import toast from "react-hot-toast";
 
 export default function EditCourier() {
@@ -10,15 +10,27 @@ export default function EditCourier() {
   const navigate = useNavigate();
   const { t } = useTranslation();
 
-  const { couriers, updateCourier } = useCourierStore();
+  const fetchCouriers = useCourierStore((s) => s.fetchCouriers);
+  const isLoading = useCourierStore((s) => s.isLoading);
+  const couriers = useCourierStore((s) => s.couriers);
+  const courier = couriers.find((item) => String(item.id) === String(id));
+  const updateCourierFromForm = useCourierStore((s) => s.updateCourierFromForm);
 
-  const courier = couriers.find((c) => String(c.id) === id);
+  useEffect(() => {
+    if (!courier && couriers.length === 0) {
+      fetchCouriers();
+    }
+  }, [courier, couriers.length, fetchCouriers]);
+
+  if (isLoading && !courier) {
+    return <div>{t("Loading...")}</div>;
+  }
 
   if (!courier) return <div>{t("Courier not found")}</div>;
 
   const handleSubmit = async (data) => {
     try {
-      await updateCourier(id, toCourierPayload(data, courier));
+      await updateCourierFromForm(id, data);
       toast.success(t("updateCourier"));
       navigate("/drivers");
     } catch (error) {
